Moved to fuzz function
